Change location and behavior of the Password Change action
This action fits better in the File menu than under Settings, so
move it there.

Also, to increase discoverability, we now show the option even if
the core is too old to support the feature. A popup message informing
the user is shown in that case.

Additionally, the changed password is now saved client-side (if the
account is setup to remember the password), thus avoiding having to
re-enter the new password on next launch.