Add support for SHA2-512 hash on Qt4 builds
Qt4 doesn't support anything higher than SHA1 in
QCryptographicHash, so it is necessary to include a fallback.  This
bundles sha512.c from PolarSSL/mbed TLS.  It is only used for Qt4
builds and does not require a dependency on external libraries that
might be hard to compile on Windows® or Mac OSX®.  It can be
removed once Qt4 is no longer supported.