Make PeerPtr work for RPC calls
While most of the functionality was already in place, it turns out
that we need to enable loading/saving PeerPtr from/to a QVariant in
order for SignalProxy to work properly.

I'm honestly not sure why it seemed to work without this when implementing
the TransferManager, but it's obvious it's required.