Provide a proper dialog for changing the core password
The password change dialog now asks for your old password, and forces
you to confirm the new password by typing it twice, as is customary.
To make this works, we need to add bidirectional communication between
client and core, forcing some more changes to the way the first iteration
of the feature was done.

Note that this commit breaks compatibility with the previous implementation.
If you run a beta1 core or client, you need to upgrade both to make things
work again. Since no real release happened in between, I don't deem it
necessary to provide backwards compat or yet another feature guard.