Add cmdline options for SSL certificate/key paths.
Add cmdline options to core to specify the SSL certificate
and SSL key path, --ssl-cert and --ssl-key respectively.
If not set, the usual path CONFIGDIR/quasselCert.pem
will be used for the cert and the key.

If only the path to the certificate is set, it is assumed
to also contain the key.