Performance tweak: rem. pointless recursive calls
For some reason there was an extra call to IrcChannel::joinIrcUser from IrcUser::joinIrcChannel. That means that in past we did:

* Call to                         IrcChannel::joinIrcUsers(...)
* From there we call              IrcUser::joinIrcChannel(this)
* And from there we called        IrcChannel::joinIrcUser(this)
* Which again called              IrcChannel::joinIrcUsers(list that contained only this 1 user)
* Which figured out this user is already in this channel and exits

That is a pointless overhead, so now we call the second function with extra parameter that tells it not to call joinIrcUser because we know this user doesn't need to be added there, thus saves few thousands of pointless function calls and user list contructions, especially while joining large channels.