Simplify MainPage and add Connect to core button in clients
Instead of overriding the paint function simply use a layout
and a QLabel in MainPage, this also enables us to add a button
to open the connect to core dialog.

This commit also hides the Nick list in a disconnected state,
to make sure the Quassel logo is centred and to keep it in line
with the All Chats view.