Add strict-oidentd mode
This mode is disabled by default.

When specified, this restricts authoratative ident responses to be either
the authuser account name, or custom mappings stored in the database.

No interface for editing these entries is currently exposed.

This increments the database version, adding a new table to store the
userid -> sysident mappings.