Add CapNegotiation Feature flag, fix CMakeLists
Add CapNegotiation Feature flag to Quassel's CoreFeatures list.
Technically, this isn't needed - newer clients can connect to older
cores just fine.  However, it may be useful to know if the core
supports cap negotiation or account-tracking in the future, and
there's no way to retroactively add this flag once 0.13 is released.

Remove an unneccessary semicolon at the end of Q_DECLARE_FLAGS.

Add quassel.h to CMakeLists.txt, so automoc includes it and it shows
up in Qt Creator.