Show account information in nick tooltips
Add row for account in tooltip based on IrcUser->account().  If not
supported, it's ignored.  Otherwise, will either show the account
name if logged in, or "Not logged in" if known to be logged out.

When account() information is available, hide the Whois account reply
as it's redundant, and might be out-of-date.