Allow client-side channel buffer merges, and associated changes.
networkmodel.cpp: Mark ChannelBufferItems as being a drop target, to allow for channel buffer merging.

bufferviewfilter.cpp: Rework BufferViewFilter::flags() to make it more clear what's going on, and remove the restriction that made QueryBuffers the only merge-able buffers.

bufferview.cpp: Rework BufferView::dropEvent() to add a bunch of comments, and allow ChannelBuffers to be merged as well as QueryBuffers. ChannelBuffers can only be the source for a merge if they are not currently joined, to prevent UI weirdness of being in a channel but not having any associated UI elements.