Make sender brackets configurable, on by default
Add a toggle to Settings -> Interface -> Chat View so sender brackets
can be enabled and disabled as desired.  Keeping with the timestamp
change, brackets are disabled by default but can be easily reenabled.

Revert Chat Monitor to include ':' as a separator between the
network:buffer:sender line, for otherwise it can be hard to read.