Hide AutoWHO for nicks without IrcUser objects
Don't require an IrcUser to exist when checking if WHO replies should
be hidden.  The nickname is enough for checking.

This mimics CoreSessionEventProcessor::processIrcEvent403, where
ERR_NOSUCHCHANNEL replies are hidden if nickname is in the list
regardless of whether or not an IrcUser object exists.

Potentially fixes AutoWHO replies showing in the server buffer when
nicks quickly join and leave channels, resulting in the IrcUser
object getting destroyed before the server replies to the AutoWHO
request.

It's difficult to recreate the issue on demand, so it's quite
possible there's more to fix.