Detect and label changes from ourself
On the core, detect when changes apply to our own nick and flag the
message as Self.

On the client, work around old nick change messages lacking the Self
flag.  Working around the other messages would be more difficult.

This fixes nick change styling not correctly applying for
"sender=self" in themes.