Add capAvailable to Network akin to capEnabled
Add capAvailable to Network, checking if the server advertised the
given capability.  This mimics capEnabled, but applies regardless of
whether or not Quassel requested the capability.

This is useful for showing elsewhere whether or not a certain
capability could've been requested.

Use this in Network::saslMaybeSupports to avoid falsely claiming SASL
support when the SASL capability wasn't advertised.  This matches the
behavior of CoreNetwork.