Encourage SASL over NickServ when server supports
Move Use SASL Authentication above Auto Identify.

Change the label for SASL according to whether or not SASL support
is advertised by the network.  If disconnected or unknown, no change.
Otherwise, set to "Use SASL Authentication (preferred)" when known,
or "Use SASL Authentication (might not work)" when not advertised.

Unfortunately there's no way to be sure, but this seems to do the
right thing.  One can still enable and disable as before.

Add explanatory tooltips to authentication options, including
recommending SASL if you need to identify before joining channels
(this gets asked a lot in #quassel).