core: connectToIrc: Prevent DNS leaks on connection when using proxy
96fe7e713 introduced a call to QHostInfo::fromName() in order to avoid
problems related to the caching of DNS queries. It however is also called
when the user has specified that a proxy should be used. This behavior
therefore introduces a DNS leak.

This commit ensures that QHostInfo::fromName() is not called when
a proxy has been specified.

(cherry picked from commit 8aa70869402f71328d1a5f92fa36aa9579dc333b)