client: Show when End Session unavailable, polish
Show when the End Session option is unavailable, adding a tooltip to
either point out an outdated core, or a client that doesn't support
it.

Before RemoteDisconnect was added to the core, it didn't share any
client features with other connected clients, so there's no way to
tell if a client supports a feature or not.

See 78b37c1a0610d9a4fb26ba0dec7337d6f7960041
"Transmit feature flag for remote disconnecting with peer info"

Disable and rename the End Session button when clicked, giving
immediate feedback when Quassel's used with a high-latency connection.