Don't terminate the core due to backend issues while setting up
If the storage backend selected by the user during core setup cannot
be initialized, the problem shall be communicated to the client and
the user can start over; the core shall not exit in this case.
Ensure that failure to initialize the selected backend only terminates
the core if it's not being setup.