common: Apply RegEx, CS column to sender, channel
Modify scopeMatch to optionally treat the rule as a regular
expression, and to treat it as case-sensitive or not.

Modify highlight rule processing to...
* Use scopeMatch for "Sender", allowing easy multi-sender rules
* Treat both "Sender" and "Channel" as regular expressions if "RegEx"
  is checked, allowing for full regular expression power.
  (Breaks existing rules with RegEx matching, but in a different way)

This means it is no longer possible to use RegEx matching for the
"Sender" and "Channel" yet not on the phrase itself.  However, it
should be easy enough to work around this by using the same phrase
regular expression Quassel does.

This should simplify common highlight rules without limiting those
who know regular expressions.  Win-win, hopefully?

Update tooltips to reflect these changes.