mono: Show a popup during database migration
Database upgrades may take a long while, during which the monolightic
client can not be used. Show a modal popup during migration to both
inform the user about what's going on, and preventing them from
interacting with the client until the internal core is up and running.

Show an "Initializing..." message in the status bar while the core
is starting up, so there's some indication for the user besides the
modal popup.