qtui: Fix icons in the AboutDlg
With the icon reorganization, no aliases are created inside resource
files anymore. Explicitly put the Breeze and Oxygen Quassel pixmaps
for the AboutDlg into pics/ and add it to the :pics/ resource.

Remove pics/quassel.png which was shipped, but not actually used
due to it missing from the resource file.