mono: "Import Legacy" for import local highlights
Rename "Import Local" as "Import Legacy", and replace references to
"Local Highlights" with "Legacy Highlights" in core-side highlights
when running in monolithic mode.

This reduces confusion over where highlights are imported when
migrating highlights in Monolithic mode.

This continues clarification efforts in a previous commit,
0182f0e3977026de65be1edb6ca1e8d7ea0184fd

Only import if question message box result is Yes, instead of not No.
This reduces the risk of accidentally importing if the message box is
ever changed to give more options or if it returns other responses.