client: Fix SettingsPage::hasChanged() failure
Fix SettingsPage::hasChanged() to compare with the correct property,
"storedValue", instead of "StoredValue".  This resulted in the
settings page sometimes errantly marking values as not changed when
they were, or vice-versa.

Some settings pages could probably be cleaned up now that this is
fixed.  That is left for future cleanup efforts.