client: Clean up chatviewsettingspage.ui
Clean up the HTML spew from the Rich Text editor for
senderPrefixComboBox, using handcrafted Qt markup to simplify it for
translators.  (Hopefully.)

CAUTION: Editing the tooltip in Qt's Rich Text editor will convert it
back to HTML.  Use the Source view of the editor.

Use <number> instead of <bool> for custom enumeration values.  This
stops Qt Creator from converting it to "false" when loading the file.