client: Poll legacy CoreInfo when dialog is open
On legacy cores (lacking SyncedCoreInfo), when CoreInfoDlg is open,
every 15 seconds resynchronize CoreInfo object to update the
"Connected Clients" count.  This matches expectations set by the
counting uptime clock, and shouldn't be a significant data impact as
info is only polled while the dialog is open.

Modern cores (with SyncedCoreInfo) will not poll, using signals to
update immediately as before.