common: Make CommitDate Unix epoch, handle legacy
Leave CommitDate (BuildDate) in Unix epoch format, converting it on
demand when displayed.  This allows clients to localize the time
shown.

Add tryFormatUnixEpoch, which checks a string for Unix epoch, parsing
as date/time if possible, otherwise letting the string through.  This
avoids breaking protocol or any existing packaging scripts that set
commit date to a string.

Check for empty commit/build dates when using them, replacing with
"Unknown date" or "unknown" as appropriate.  Don't encode this in
Quassel::setupBuildInfo() to allow for location-specific treatment
and client-side translation.