quassel: Install signal handlers in init() rather than ctor
Unix signal handlers require e.g. command line options, which are
not available at construction time of the Quassel instance.
Move signal handler registration into init().

Add TODO comments to remind us of the fact that we're not supposed
to use unsafe methods in signal handlers (which we currently do),
and should rework signal handling to be safe.