common: Empty highlight name matches all messages
Don't force core-side highlight rules to have non-empty names,
allowing for matching all messages by leaving the name empty.

Requires core and client support, but as this has not been in a
released version, and it's near release, backwards-compatibility
shouldn't be required.

Update tooltips to note this.  (After cleanup, only a single change!)

This wasn't easily possible before as highlight names were the
identifier for the highlight rule.  Now that we have IDs for
highlight rules, there's no need to force names to be non-empty.

NOTE:  This could cause confusion, but the flexibility in not having
to use RegEx for the sake of matching every message seems worth the
risk of people unintentionally setting empty highlight rules that
match every message.

Local highlights are not changed as they don't have IDs.